@@ -1,7 +1,9 @@
+#if canImport(UIKit)
 import Foundation
 
-public protocol TextDecorator: TokenTraitProvider, Hashable {
+public protocol TextDecorator: TokenTraitProvider, Hashable, Sendable {
 
+    @MainActor
     func decorate(
         typography: TypographyValue,
         context: ComponentContext
@@ -13,6 +15,7 @@ extension TextDecorator where
     Input == TypographyValue,
     Output == TypographyValue {
 
+    @MainActor
     public func decorate(typography: TypographyValue, context: ComponentContext) -> TypographyValue {
         decorate(typography, theme: context.tokenTheme)
     }
@@ -40,3 +43,4 @@ extension Array: TextDecorator where Element: TextDecorator {
         }
     }
 }
+#endif
