@@ -1,3 +1,4 @@
+#if canImport(UIKit)
 import SwiftUI
 
 /// Общий протокол для всех UI-компонентов.
@@ -111,6 +112,9 @@ public protocol Component: View, Equatable {
     ///
     /// - SeeAlso: ``ComponentSizing``
     /// - SeeAlso: ``ComponentContext``
+    #if swift(<6.0)
+    @MainActor
+    #endif
     func sizing(
         fitting size: CGSize,
         context: ComponentContext
@@ -119,7 +123,7 @@ public protocol Component: View, Equatable {
 
 extension Component {
 
-    internal func isEqual(to other: any Component) -> Bool {
+    internal nonisolated func isEqual(to other: any Component) -> Bool {
         guard let other = other as? Self else {
             return false
         }
@@ -127,3 +131,4 @@ extension Component {
         return self == other
     }
 }
+#endif
