@@ -1,3 +1,4 @@
+#if canImport(UIKit)
 import SwiftUI
 
 /// Контекст компонента.
@@ -89,12 +90,11 @@ public struct ComponentContext {
     internal let environment: EnvironmentValues?
     internal let overrides: [PartialKeyPath<EnvironmentValues>: ComponentContextOverride]
 
+    @MainActor
     internal func resolveValue<Value>(at keyPath: KeyPath<EnvironmentValues, Value>) -> Value {
-        if let value = overrides[keyPath].flatMap({ $0.value as? Value }) {
-            return value
-        }
-
-        let environment = environment ?? .default
+        let environment = overrides
+            .values
+            .reduce(into: environment ?? .default) { $1.override(for: &$0) }
 
         return environment[keyPath: keyPath]
     }
@@ -122,6 +122,7 @@ public struct ComponentContext {
     ///   - keyPath: Ключ переменной окружения.
     ///   - transform: Замыкание, трасформирующее текущее значение переменной окружения.
     /// - Returns: Окружение с переопределенной переменной.
+    @MainActor
     public func transformEnvironment<Value>(
         _ keyPath: WritableKeyPath<EnvironmentValues, Value>,
         transform: @escaping (inout Value) -> Void
@@ -139,6 +140,7 @@ public struct ComponentContext {
     ///   - keyPath: Ключ переменной окружения.
     ///   - value: Новое значение переменной.
     /// - Returns: Окружение с переопределенной переменной.
+    @MainActor
     public func environment<Value>(
         _ keyPath: WritableKeyPath<EnvironmentValues, Value>,
         _ value: Value
@@ -153,6 +155,7 @@ public struct ComponentContext {
     ///
     /// - Parameter isDisabled: новое значение.
     /// - Returns: Окружение с переопределенной переменной.
+    @MainActor
     public func disabled(_ isDisabled: Bool = true) -> Self {
         self.isEnabled(!isDisabled && self.isEnabled)
     }
@@ -161,6 +164,7 @@ public struct ComponentContext {
     ///
     /// - Parameter keyPath: Ключ переменной окружения.
     /// - Returns: Значение переменной окружения.
+    @MainActor
     public subscript<Value>(dynamicMember keyPath: KeyPath<EnvironmentValues, Value>) -> Value {
         resolveValue(at: keyPath)
     }
@@ -185,8 +189,11 @@ extension ComponentContext {
     ///
     /// Рекомендуется использовать в корневом UIKit-компоненте.
     /// Но также допускается использование по месту для обнуления контекста или в целях миграции.
-    public static let `default` = Self(
-        environment: nil,
-        overrides: [:]
-    )
+    public static var `default`: ComponentContext {
+        Self(
+            environment: nil,
+            overrides: [:]
+        )
+    }
 }
+#endif
