@@ -1,14 +1,14 @@
+#if canImport(UIKit)
 import SwiftUI
-import UIKit
 
 internal struct ComponentContainerSizeKey: EnvironmentKey {
 
-    internal static let defaultValue = UIScreen.main.bounds.size
+    internal static let defaultValue: CGSize? = nil
 }
 
 extension EnvironmentValues {
 
-    /// Размеры ближайшего известного контейнера .
+    /// Размеры ближайшего известного контейнера.
     ///
     /// Используется в качестве ограничивающих размеров для определения размеров самих компонентов
     /// при встраивании UIKit-компонента в SwiftUI-представление.
@@ -17,8 +17,9 @@ extension EnvironmentValues {
     ///
     /// - Note: Нет необходимости самостоятельно устанавливать значение для этой переменной,
     ///         его переопределяют встроенные компоненты.
-    public var componentContainerSize: CGSize {
+    public var componentContainerSize: CGSize? {
         get { self[ComponentContainerSizeKey.self] }
         set { self[ComponentContainerSizeKey.self] = newValue }
     }
 }
+#endif
