@@ -1,8 +1,9 @@
+#if canImport(UIKit)
 import SwiftUI
 
 internal struct ComponentLayoutInvalidationEnvironmentKey: EnvironmentKey {
 
-    internal static let defaultValue = { }
+    internal static let defaultValue = { @MainActor in }
 }
 
 extension EnvironmentValues {
@@ -20,7 +21,7 @@ extension EnvironmentValues {
     /// - Warning: Не рекомендуется вызывать инвалидацию при изменении внешнего состояния,
     ///            переданного через байндинги или через ручное связывание замыканиями.
     ///            Компонент должен выполнять это действие только, если изменилось его собственное состояние.
-    public internal(set) var invalidateComponentLayout: () -> Void {
+    public internal(set) var invalidateComponentLayout: @MainActor () -> Void {
         get { self[ComponentLayoutInvalidationEnvironmentKey.self] }
         set { self[ComponentLayoutInvalidationEnvironmentKey.self] = newValue }
     }
@@ -32,9 +33,10 @@ extension ComponentContext {
     ///
     /// Дополнительное действие будет выполнено до выполнения уже имеющихся действий инвалидации.
     ///
-    /// - Parameter layoutInvalidation: Дополнительное действие для инвалидации лэйаута.
+    /// - Parameter invalidation: Дополнительное действие для инвалидации лэйаута.
     /// - Returns: Окружение с добавленным действием для инвалидации лэйаута.
-    public func componentLayoutInvalidation(_ invalidation: @escaping () -> Void) -> Self {
+    @MainActor
+    public func componentLayoutInvalidation(_ invalidation: @escaping @MainActor () -> Void) -> Self {
         transformEnvironment(\.invalidateComponentLayout) { currentInvalidation in
             currentInvalidation = { [currentInvalidation] in
                 invalidation()
@@ -43,3 +45,4 @@ extension ComponentContext {
         }
     }
 }
+#endif
