@@ -1,3 +1,4 @@
+#if canImport(UIKit)
 import UIKit
 
 /// UIKit-представление компонента-обертки со стертым типом.
@@ -21,8 +22,7 @@ public final class AnyManualComponentView: UIView {
         wrappedView.isFirstResponder
     }
 
-    // swiftlint:disable:next explicit_acl
-    override init(frame: CGRect) {
+    public override init(frame: CGRect) {
         super.init(frame: frame)
 
         addSubview(wrappedView)
@@ -51,8 +51,7 @@ public final class AnyManualComponentView: UIView {
     }
 
     @available(*, unavailable)
-    // swiftlint:disable:next explicit_acl
-    required init?(coder: NSCoder) {
+    public required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
 
@@ -83,3 +82,4 @@ extension AnyManualComponentView: ComponentView {
         invalidateIntrinsicContentSize()
     }
 }
+#endif
