@@ -1,11 +1,13 @@
+#if canImport(UIKit)
 import Foundation
 
 extension BackgroundColorDecorator: TokenTraitProvider where Value == TypographyValue { }
 extension BackgroundColorDecorator: TextDecorator where Value == TypographyValue { }
 
-extension Text2 {
+extension Text {
 
     public func backgroundColor(_ backgroundColor: ColorToken?) -> Self {
         decorated(by: BackgroundColorDecorator(backgroundColor: backgroundColor))
     }
 }
+#endif
