@@ -1,11 +1,13 @@
+#if canImport(UIKit)
 import Foundation
 
 extension StrikethroughDecorator: TokenTraitProvider where Value == TypographyValue { }
 extension StrikethroughDecorator: TextDecorator where Value == TypographyValue { }
 
-extension Text2 {
+extension Text {
 
     public func strikethrough(_ strikethrough: TypographyLineToken? = .single) -> Self {
         decorated(by: StrikethroughDecorator(strikethrough: strikethrough))
     }
 }
+#endif
