@@ -1,23 +1,40 @@
+#if canImport(UIKit)
 import UIKit
 
 public final class TextView: UILabel {
 
-    private var content: Text2?
+    private var content: Text?
     private var context: ComponentContext?
 
     private var partThresholds: [Int] = []
 
     private var hoveredPartIndex: Int?
     private var pressedPartIndex: Int?
-    private var animatedPartIndex: Int?
 
     private var hasTapAction = false
+
+    private lazy var operationQueue: TextOperationQueue = {
+        TextOperationQueue(layer: layer)
+    }()
 
     public override init(frame: CGRect) {
         super.init(frame: frame)
 
+        #if os(iOS)
+        let hoverGestureRecognizer = UIHoverGestureRecognizer(
+            target: self,
+            action: #selector(onHoverGesture(recognizer:))
+        )
+
+        addGestureRecognizer(hoverGestureRecognizer)
+        #endif
+
         tokens.customBinding { view, theme in
-            view.updateAttributedText(theme: theme)
+            view.updateAttributedText(
+                hoveredPartIndex: view.hoveredPartIndex,
+                pressedPartIndex: view.pressedPartIndex,
+                theme: theme
+            )
         }
     }
 
@@ -26,7 +43,7 @@ public final class TextView: UILabel {
         fatalError("init(coder:) has not been implemented")
     }
 
-    private func updateAttributedText(theme: TokenTheme) {
+    private func updateAttributedText(hoveredPartIndex: Int?, pressedPartIndex: Int?, theme: TokenTheme) {
         let context = context?
             .tokenThemeKey(theme.key)
             .tokenThemeScheme(theme.scheme)
@@ -40,51 +57,70 @@ public final class TextView: UILabel {
         partThresholds.removeAll(keepingCapacity: true)
 
         for (partIndex, part) in content.parts.enumerated() {
-            let partState = TextPartState(
-                isHovered: partIndex == hoveredPartIndex,
-                isPressed: partIndex == pressedPartIndex
-            )
+            let partContext = context
+                .isTextPartHovered(partIndex == hoveredPartIndex)
+                .isTextPartPressed(partIndex == pressedPartIndex)
 
-            let partAttributedText = part.attributedText(
-                context: context.textPartState(partState)
-            )
+            let partAttributedText = part.attributedText(context: partContext)
 
             attributedText.append(partAttributedText)
             partThresholds.append(attributedText.length)
         }
 
         self.attributedText = attributedText
+
+        numberOfLines = context.lineLimit ?? .zero
+        lineBreakMode = content.lineBreakMode
+    }
+
+    private func updateAttributedText(hoveredPartIndex: Int?, pressedPartIndex: Int?) {
+        updateAttributedText(
+            hoveredPartIndex: hoveredPartIndex,
+            pressedPartIndex: pressedPartIndex,
+            theme: tokens.theme
+        )
     }
 
     private func updateAttributedText() {
-        updateAttributedText(theme: tokens.theme)
+        let hoveredPartIndex = hoveredPartIndex
+        let pressedPartIndex = pressedPartIndex
+
+        let animation = content?.animation ?? context?.textAnimation
+
+        let updateAnimation = attributedText != nil && window != nil
+            ? animation?.update
+            : nil
+
+        operationQueue.addOperation(animation: updateAnimation) {
+            self.updateAttributedText(
+                hoveredPartIndex: hoveredPartIndex,
+                pressedPartIndex: pressedPartIndex
+            )
+        }
     }
 
-    private func animatePressedPart() {
-        animatedPartIndex = pressedPartIndex
+    private func updateHoveredPart(index: Int?) {
+        guard hoveredPartIndex != index else {
+            return
+        }
+
+        hoveredPartIndex = index
+
+        let hoveredPartIndex = hoveredPartIndex
+        let pressedPartIndex = pressedPartIndex
 
         let animation = content?.animation ?? context?.textAnimation
 
         let partAnimation = pressedPartIndex == nil
-            ? animation?.unpress
-            : animation?.press
+            ? (window == nil ? nil : animation?.unhover)
+            : (window == nil ? nil : animation?.hover)
 
-        guard let partAnimation else {
-            return updateAttributedText()
+        operationQueue.addOperation(animation: partAnimation) {
+            self.updateAttributedText(
+                hoveredPartIndex: hoveredPartIndex,
+                pressedPartIndex: pressedPartIndex
+            )
         }
-
-        let transition = partAnimation
-            .caTransition
-            .resolve(for: tokens.theme)
-
-        transition.delegate = self
-
-        layer.add(
-            transition,
-            forKey: "text"
-        )
-
-        updateAttributedText()
     }
 
     private func updatePressedPart(index: Int?) {
@@ -94,18 +130,24 @@ public final class TextView: UILabel {
 
         pressedPartIndex = index
 
-        let isAnimating = layer.animationKeys()?.contains { key in
-            layer.animation(forKey: key)?.delegate === self
-        } ?? false
+        let hoveredPartIndex = hoveredPartIndex
+        let pressedPartIndex = pressedPartIndex
 
-        if isAnimating {
-            updateAttributedText()
-        } else {
-            animatePressedPart()
+        let animation = content?.animation ?? context?.textAnimation
+
+        let partAnimation = pressedPartIndex == nil
+            ? (window == nil ? nil : animation?.unpress)
+            : (window == nil ? nil : animation?.press)
+
+        operationQueue.addOperation(animation: partAnimation) {
+            self.updateAttributedText(
+                hoveredPartIndex: hoveredPartIndex,
+                pressedPartIndex: pressedPartIndex
+            )
         }
     }
 
-    private func touchCharacterIndex(touchLocation: CGPoint) -> Int? {
+    private func gestureCharacterIndex(location: CGPoint) -> Int? {
         guard let attributedText else {
             return nil
         }
@@ -130,26 +172,24 @@ public final class TextView: UILabel {
             y: 0.5 * (size.height - textBounds.size.height) - textBounds.origin.y
         )
 
-        let touchTextLocation = CGPoint(
-            x: touchLocation.x - textOffset.x,
-            y: touchLocation.y - textOffset.y
+        let gestureTextLocation = CGPoint(
+            x: location.x - textOffset.x,
+            y: location.y - textOffset.y
         )
 
         return layoutManager.characterIndex(
-            for: touchTextLocation,
+            for: gestureTextLocation,
             in: textContainer,
             fractionOfDistanceBetweenInsertionPoints: nil
         )
     }
 
-    private func touchPartIndex(touch: UITouch) -> Int? {
-        let touchLocation = touch.location(in: self)
-
-        guard bounds.contains(touchLocation) else {
+    private func gesturePartIndex(location: CGPoint) -> Int? {
+        guard bounds.contains(location) else {
             return nil
         }
 
-        guard let characterIndex = touchCharacterIndex(touchLocation: touchLocation) else {
+        guard let characterIndex = gestureCharacterIndex(location: location) else {
             return nil
         }
 
@@ -162,6 +202,22 @@ public final class TextView: UILabel {
             : nil
     }
 
+    #if os(iOS)
+    @objc
+    private func onHoverGesture(recognizer: UIHoverGestureRecognizer) {
+        switch recognizer.state {
+        case .began, .changed:
+            updateHoveredPart(index: gesturePartIndex(location: recognizer.location(in: self)))
+
+        case .ended, .cancelled, .failed, .possible:
+            updateHoveredPart(index: nil)
+
+        @unknown default:
+            updateHoveredPart(index: nil)
+        }
+    }
+    #endif
+
     public override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
         super.traitCollectionDidChange(previousTraitCollection)
 
@@ -169,7 +225,9 @@ public final class TextView: UILabel {
         let contentSizeCategory = traitCollection.preferredContentSizeCategory
 
         if previousContentSizeCategory != contentSizeCategory {
-            updateAttributedText()
+            if !operationQueue.isRunning {
+                updateAttributedText()
+            }
         }
     }
 
@@ -180,7 +238,7 @@ public final class TextView: UILabel {
             return
         }
 
-        updatePressedPart(index: touchPartIndex(touch: touch))
+        updatePressedPart(index: gesturePartIndex(location: touch.location(in: self)))
     }
 
     public override func touchesMoved(_ touches: Set<UITouch>, with event: UIEvent?) {
@@ -190,7 +248,7 @@ public final class TextView: UILabel {
             return updatePressedPart(index: nil)
         }
 
-        updatePressedPart(index: touchPartIndex(touch: touch))
+        updatePressedPart(index: gesturePartIndex(location: touch.location(in: self)))
     }
 
     public override func touchesCancelled(_ touches: Set<UITouch>, with event: UIEvent?) {
@@ -222,7 +280,7 @@ public final class TextView: UILabel {
 
         isUserInteractionEnabled = false
 
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) {
+        operationQueue.addOperation {
             self.isUserInteractionEnabled = self.context?.isEnabled ?? true
 
             partTapAction?()
@@ -231,32 +289,26 @@ public final class TextView: UILabel {
     }
 }
 
-extension TextView: CAAnimationDelegate {
-
-    public func animationDidStop(_ animation: CAAnimation, finished: Bool) {
-        if animatedPartIndex != pressedPartIndex {
-            animatePressedPart()
-        }
-    }
-}
-
 extension TextView: FallbackManualComponentView {
 
-    public static func sizing(
-        for content: Text2,
+    public static func size(
+        for content: Text,
         fitting size: CGSize,
         context: ComponentContext
-    ) -> ComponentSizing {
-        let size = Self.size(
+    ) -> CGSize {
+        let attributedText = attributedText(
             for: content,
-            fitting: size,
             context: context
         )
 
-        return ComponentSizing(size: size)
+        return attributedText.size(
+            fitting: size,
+            lineLimit: content.lineLimit,
+            lineBreakMode: content.lineBreakMode
+        )
     }
 
-    public func update(with content: Text2, context: ComponentContext) {
+    public func update(with content: Text, context: ComponentContext) {
         let typography = content.typography ?? context.textTypography
 
         let decoration = context
@@ -275,9 +327,10 @@ extension TextView: FallbackManualComponentView {
         self.content = content
         self.context = context
 
+        isUserInteractionEnabled = context.isEnabled
+
         hoveredPartIndex = nil
         pressedPartIndex = nil
-        animatedPartIndex = nil
 
         if content.tapAction == nil {
             hasTapAction = content
@@ -289,16 +342,14 @@ extension TextView: FallbackManualComponentView {
 
         updateAttributedText()
 
-        isUserInteractionEnabled = context.isEnabled
-        numberOfLines = context.lineLimit ?? .zero
-        lineBreakMode = content.lineBreakMode
+        sizeToFit()
     }
 }
 
 extension TextView {
 
     public static func attributedText(
-        for content: Text2,
+        for content: Text,
         context: ComponentContext
     ) -> NSAttributedString {
         let typography = content.typography ?? context.textTypography
@@ -320,7 +371,7 @@ extension TextView {
 
         for part in content.parts {
             let partAttributedText = part.attributedText(
-                context: context.textPartState(.default)
+                context: context
             )
 
             attributedText.append(partAttributedText)
@@ -328,21 +379,5 @@ extension TextView {
 
         return attributedText
     }
-
-    public static func size(
-        for content: Text2,
-        fitting size: CGSize,
-        context: ComponentContext
-    ) -> CGSize {
-        let attributedText = attributedText(
-            for: content,
-            context: context
-        )
-
-        return attributedText.size(
-            fitting: size,
-            lineLimit: context.lineLimit,
-            lineBreakMode: content.lineBreakMode
-        )
-    }
 }
+#endif
