@@ -1,3 +1,5 @@
+#if canImport(UIKit)
+import Combine
 import UIKit
 
 extension UIWindowScene: TokenView {
@@ -35,19 +37,25 @@ extension UIWindowScene: TokenView {
     }
 }
 
-private var tokenViewWindowScenesObservation: NSObjectProtocol?
+@MainActor
+private var tokenViewWindowScenesObservation: AnyCancellable?
 
 extension UIWindowScene {
 
     internal static func handleTokenViewEvents() {
-        tokenViewWindowScenesObservation = NotificationCenter.default.addObserver(
-            forName: UIWindowScene.didActivateNotification,
-            object: nil,
-            queue: nil
-        ) { notification in
-            if let tokenView = notification.object as? TokenView {
+        tokenViewWindowScenesObservation = NotificationCenter.default.publisher(
+            for: UIWindowScene.didActivateNotification
+        )
+        .compactMap { $0.object as? TokenView }
+        .sink { tokenView in
+            if Thread.isMainThread {
                 tokenView.tokenViewManager.updateTheme()
+            } else {
+                Task { @MainActor in
+                    tokenView.tokenViewManager.updateTheme()
+                }
             }
         }
     }
 }
+#endif
