@@ -1,3 +1,4 @@
+#if canImport(UIKit)
 import UIKit
 
 extension TokenViewProperties where View: UIControl {
@@ -16,9 +17,10 @@ extension TokenViewProperties where View: UIControl {
     }
 }
 
-extension UIControl.State: Hashable {
+extension UIControl.State: @retroactive Hashable {
 
     public func hash(into hasher: inout Hasher) {
         hasher.combine(rawValue)
     }
 }
+#endif
