@@ -1,3 +1,4 @@
+#if canImport(UIKit)
 import UIKit
 
 extension TokenViewProperties where View: UITextField {
@@ -37,11 +38,5 @@ extension TokenViewProperties where View: UITextField {
             view.disabledBackground = value?.uiImage
         }
     }
-
-// TODO: implement for text
-//    public var attributedPlaceholder: TokenViewProperty<TokenText, Void> {
-//        property { view, value, theme in
-//            view.attributedPlaceholder = value?.nsAttributedString(for: theme)
-//        }
-//    }
 }
+#endif
