@@ -1,4 +1,4 @@
-import Foundation
+import CoreGraphics
 
 public typealias AnimationToken = Token<AnimationValue>
 
