@@ -1,8 +1,8 @@
 import CoreGraphics
 
-extension CGPoint: TokenValue { }
+extension CGPoint: @retroactive @unchecked Sendable, TokenValue { }
 
-extension CGPoint: Hashable {
+extension CGPoint: @retroactive Hashable {
 
     public func hash(into hasher: inout Hasher) {
         hasher.combine(x)
