@@ -1,6 +1,7 @@
 import CoreGraphics
 
 extension CGRect:
+    @retroactive @unchecked Sendable,
     TokenValue,
     DecorableByInsets,
     DecorableByOutsets {
@@ -24,7 +25,7 @@ extension CGRect:
     }
 }
 
-extension CGRect: Hashable {
+extension CGRect: @retroactive Hashable {
 
     public func hash(into hasher: inout Hasher) {
         hasher.combine(origin)
