@@ -1,10 +1,15 @@
+#if canImport(UIKit)
 import UIKit
+#endif
+
 import SwiftUI
 
 public enum ImageSource: TokenTraitProvider, Sendable {
 
-    case uiImage(UIImage)
     case resource(name: String, bundle: Bundle)
+
+    #if canImport(UIKit)
+    case uiImage(UIImage)
 
     public var uiImage: UIImage {
         switch self {
@@ -19,14 +24,18 @@ public enum ImageSource: TokenTraitProvider, Sendable {
             return uiImage
         }
     }
+    #endif
 
     public var image: Image {
         switch self {
-        case let .uiImage(uiImage):
-            return Image(uiImage: uiImage)
-
         case let .resource(name, bundle):
             return Image(name, bundle: bundle)
+
+        #if canImport(UIKit)
+
+        case let .uiImage(uiImage):
+            return Image(uiImage: uiImage)
+        #endif
         }
     }
 }
@@ -35,14 +44,17 @@ extension ImageSource: Equatable {
 
     public static func == (lhs: Self, rhs: Self) -> Bool {
         switch (lhs, rhs) {
-        case let (.uiImage(lhs), .uiImage(rhs)):
-            return lhs.isEqual(rhs)
-
         case let (.resource(lhsName, lhsBundle), .resource(rhsName, rhsBundle)):
             return (lhsName == rhsName) && (lhsBundle == rhsBundle)
 
+        #if canImport(UIKit)
+
+        case let (.uiImage(lhs), .uiImage(rhs)):
+            return lhs.isEqual(rhs)
+
         default:
             return false
+        #endif
         }
     }
 }
@@ -51,12 +63,15 @@ extension ImageSource: Hashable {
 
     public func hash(into hasher: inout Hasher) {
         switch self {
-        case let .uiImage(uiImage):
-            hasher.combine(uiImage)
-
         case let .resource(name, bundle):
             hasher.combine(name)
             hasher.combine(bundle)
+
+        #if canImport(UIKit)
+
+        case let .uiImage(uiImage):
+            hasher.combine(uiImage)
+        #endif
         }
     }
 }
