@@ -1,12 +1,13 @@
+#if canImport(UIKit)
 import UIKit
 
 public struct TypographyLineValue:
     TokenValue,
-    DecorableByColor,
+    Changeable,
     Sendable {
 
-    public let style: NSUnderlineStyle
-    public let color: ColorValue?
+    public var style: NSUnderlineStyle
+    public var color: ColorValue?
 
     public init(
         style: NSUnderlineStyle,
@@ -17,14 +18,7 @@ public struct TypographyLineValue:
     }
 }
 
-extension TypographyLineValue: Changeable {
-
-    public init(copy: ChangeableWrapper<Self>) {
-        self.init(
-            style: copy.style,
-            color: copy.color
-        )
-    }
+extension TypographyLineValue: DecorableByColor {
 
     public func color(_ color: ColorValue?) -> Self {
         changing { $0.color = color }
@@ -44,3 +38,4 @@ extension TypographyLineValue {
         )
     }
 }
+#endif
